place temporarily

+++ Bibliofilia/views.py
@@ -6,21 +6,23 @@ from Bibliofilia.forms import BibliofiliaForm
 from Bibliofilia.models import *
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import get_object_or_404, render, redirect
-from django.http import HttpResponseNotFound, HttpResponseRedirect, HttpResponse, JsonResponse
+from django.http import HttpResponseForbidden, HttpResponseNotFound, HttpResponseRedirect, HttpResponse, JsonResponse
 from django.urls import reverse
 from django.core import serializers
 from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages 
 from django.views.decorators.csrf import csrf_exempt
-# Create your views here.
+from authentication.models import UserProfile
 
 
 @login_required(login_url='auth:signin')
 def showForum(request):
-    forum = Forum.objects.annotate(replies_count=Count('replies'))
+    forum = Forum.objects.all()
+    member = UserProfile.objects.get(user=request.user)
     context = {
         'name': request.user.username,
-        'forum': forum, 
+        'forum': forum,
+        'member': member,
     }
 
     return render(request, "mainforum.html", context)
@@ -32,6 +34,7 @@ def showChildForum(request, forum_id):
             forum = Forum.objects.get(pk=forum_id)
             forum_head_msg = forum.userReview 
             forum_title = forum.BookName
+            member = UserProfile.objects.get(user=request.user)
         except Forum.DoesNotExist:
             return HttpResponseNotFound("Forum not found")
 
@@ -41,6 +44,7 @@ def showChildForum(request, forum_id):
             'name': request.user.username,
             'forum': forum,
             'forum_title':forum_title,
+            'member': member,
         }
         return render(request, "childforum.html", context)
     else:
@@ -69,14 +73,6 @@ def create_Forum(request):
     return render(request, "create_Forum.html", context)
 
 
-def show_json(request):
-    data = Forum.objects.all()
-    return HttpResponse(serializers.serialize("json", data), content_type="application/json")
-
-def show_json_by_id(request, id):
-    data = Forum.objects.filter(pk=id)
-    return HttpResponse(serializers.serialize("json", data), content_type="application/json")
-
 def register(request):
     form = UserCreationForm()
 
@@ -90,34 +86,14 @@ def register(request):
     return render(request, 'register.html', context)
 
 @csrf_exempt
-def add_amount(request, forum_id):
-    if request.method == 'POST':
-        forum = Forum.objects.get(pk=forum_id) #Mengakses Forum yang ingin dimodifikasi
-        forum.user = request.user
-        if forum.amount > 0:
-            forum.amount += 1
-            forum.save()
-        return HttpResponse(b"ADDED", status=201)
-    return HttpResponseNotFound()
-
-@csrf_exempt
-def reduce_amount(request, forum_id):
-    if request.method == 'POST':
-        forum = Forum.objects.get(pk=forum_id)
-        forum.user = request.user
-        if forum.amount > 1:
-            forum.amount -= 1
-            forum.save()
-        else:
-            forum.delete()
-        return HttpResponse(b"REDUCED", status=201)
-    return HttpResponseNotFound()
-
-@csrf_exempt
 def delete_Forum(request, forum_id):
-    print(request)
+    user = UserProfile.objects.get(user=request.user)
+    if(user.role != 'A'):
+        response = HttpResponseForbidden("Access Denied")
+        return response
+    
     if request.method == 'DELETE':
-        forum = Forum.objects.get(pk=forum_id, user=request.user)
+        forum = Forum.objects.get(pk=forum_id)
         forum.user = request.user
         forum.delete()
         return HttpResponse(b"REMOVED", status=201)
@@ -126,12 +102,33 @@ def delete_Forum(request, forum_id):
 
 @csrf_exempt
 def delete_Replies(request, reply_id):
-    print(request, reply_id)
+    user = UserProfile.objects.get(user=request.user)
+    print(request.user.username, user.role)
+    if(user.role != 'A'):
+        response = HttpResponseForbidden("Access Denied")
+        return response
+
     if request.method == 'DELETE':
-        forumReplies = ForumReply.objects.get(id=reply_id)
-        # forumReplies.user = request.user
-        forumReplies.delete()
-        return HttpResponse(b"REMOVED", status=201)
+        try:
+            # Retrieve the ForumReply
+            forumReplies = ForumReply.objects.get(pk=reply_id)
+            
+            # Retrieve the related Forum
+            forum = forumReplies.forum
+            
+            # Ensure that the user is the owner of the ForumReply
+            if user.role == 'A':
+                # Decrement the repliesTotal field of the related Forum
+                if forum.repliesTotal > 0:
+                    forum.repliesTotal -= 1
+                    forum.save()  # Save the Forum object to update repliesTotal
+                # Delete the ForumReply
+                forumReplies.delete()
+                return HttpResponse(b"REMOVED", status=201)
+            else:
+                return HttpResponse(b"Unauthorized", status=401)
+        except ForumReply.DoesNotExist:
+            return HttpResponseNotFound("ForumReply not found")
     return HttpResponseNotFound()
 
 
@@ -156,7 +153,7 @@ def add_Forum_ajax(request):
         userReview = request.POST.get("userReview")    
         user = request.user
 
-        new_product = Forum(BookName=BookName,bookPicture=bookPicture,userReview=userReview,forumsDescription=forumsDescription,repliesTotal=0, user=user)
+        new_product = Forum(BookName=BookName,bookPicture=bookPicture,userReview=userReview,forumsDescription=forumsDescription,repliesTotal=0, user=user, username=user.username)
         
         new_product.save()
